Fix bug on concurrent import
In some case when Loader#import was called concurrent (e.g. multiple
events importing same file, that was not yet loaded) a concurrent call
to Loader#get tried to execute a module before it's dependencies where
loaded. The first #import call was still stuck waiting on dependencies
but the second call found a entry in _registry and already started
executing the module. This fix lets #import wait on a ready promise if
an entry already exists otherwise it awaits on #load.